+++ src/utils.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
 const props = {
-  booleans: ['hasError', 'isValid', 'isDirty', 'isPristine', 'isDefault', 'isEmpty', 'focus', 'touched', 'changed', 'disabled'],
+  booleans: ['hasError', 'isValid', 'isDirty', 'isPristine', 'isDefault', 'isEmpty', 'focused', 'touched', 'changed', 'disabled'],
   field: ['value', 'initial', 'default', 'label', 'placeholder', 'disabled', 'related', 'options', 'bindings', 'type', 'error'],
   separated: ['values', 'initials', 'defaults', 'labels', 'placeholders', 'disabled', 'related', 'options', 'bindings', 'types'],
   function: ['onSubmit', 'observers', 'interceptors', 'parse', 'format'],
@@ -13,7 +13,7 @@ const props = {
     isDefault: 'every',
     isEmpty: 'every',
     hasError: 'some',
-    focus: 'some',
+    focused: 'some',
     touched: 'some',
     changed: 'some',
     disabled: 'every',